# api.yml
openapi: 3.0.3
info:
  title: Clinic Patient Management API
  description: API for managing users, authentication, doctors, medication, appointments, allergies, notifications, and reports.
  version: 1.0.0
servers:
  - url: https://ehrbackend-production-a3a0.up.railway.app/api
paths:
...
  /users/{id}:
    get:
      summary: Get user by ID
      tags: [Users]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: EHR ID of the user
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update user by ID
      tags: [Users]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: EHR ID of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequestDTO'
      responses:
        '200':
          description: Updated user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
        '400':
          description: Update failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete user by ID
      tags: [Users]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: EHR ID of the user
      responses:
        '200':
          description: User deleted successfully
        '400':
          description: Delete failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}/BloodGroup:
    patch:
      summary: Update user's blood group
      tags: [Users]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: EHR ID of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequestDTO'
      responses:
        '200':
          description: Updated user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
        '400':
          description: Update failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/search-patients:
    get:
      summary: Search patients
      tags: [Users]
      parameters:
        - in: query
          name: query
          schema:
            type: string
          required: true
          description: Search query string
      responses:
        '200':
          description: List of matching patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDTO'

  # --------------------------------------------------------------------
  # Doctors
  # --------------------------------------------------------------------
...
components:
  schemas:
...
    UserRequestDTO:
      type: object
      properties:
        address:
          $ref: '#/components/schemas/Address'
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        dob:
          type: string
          description: Date in mm-dd-yyyy format
          example: 07-09-1990
        password:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        phoneNo:
          type: integer
          format: int64
        bloodGroup:
          $ref: '#/components/schemas/BloodGroup'
        gender:
          $ref: '#/components/schemas/Gender'
        emergencyContact:
          $ref: '#/components/schemas/Emergency'
        doctorProfile:
          $ref: '#/components/schemas/DoctorProfile'
        maritalStatus:
          type: string
    UserResponseDTO:
      type: object
      properties:
        email:
          type: string
        ehrId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phoneNo:
          type: integer
          format: int64
        dob:
          type: string
          description: Date in mm-dd-yyyy format
          example: 07-09-1990
        gender:
          $ref: '#/components/schemas/Gender'
        role:
          $ref: '#/components/schemas/Role'
        bloodGroup:
          $ref: '#/components/schemas/BloodGroup'
        address:
          $ref: '#/components/schemas/Address'
        doctorProfile:
          $ref: '#/components/schemas/DoctorProfile'
        maritalStatus:
          type: string
        emergencyContact:
          $ref: '#/components/schemas/Emergency'
    Address:
      type: object
      properties:
        addressLine:
          type: string
        state:
          type: string
        city:
          type: string
        pin_code:
          type: integer
          format: int64
          description: Postal code (length 6)
    Emergency:
      type: object
      properties:
        contactName:
          type: string
        contactPhoneNo:
          type: integer
          format: int64
          description: Contact phone number (length 10)
        relationship:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer

    # -----------------------------------------------------------------
    # Doctors
    # -----------------------------------------------------------------
...
    DoctorProfile:
      type: object
      properties:
        aadhaarNumber:
          type: string
        panNumber:
          type: string
        specialization:
          type: string
        licenseNumber:
          type: string
        aboutDoctor:
          type: string
        degrees:
          type: array
          items:
            type: string
...
    Role:
      type: string
      enum: [PATIENT, DOCTOR, ADMIN]
    Gender:
      type: string
      enum: [MALE, FEMALE, Non_binary, Prefer_not_to_say]
    BloodGroup:
      type: string
      enum: [A_POSITIVE, A_NEGATIVE, B_POSITIVE, B_NEGATIVE, AB_POSITIVE, AB_NEGATIVE, O_POSITIVE, O_NEGATIVE]
...